rating model functions added

+++ database/migrations/2019_10_30_131426_create_application_ratings_table.php
@@ -17,7 +17,7 @@ class CreateApplicationRatingsTable extends Migration
             $table->increments('id');
             $table->unsignedInteger('job_post_id');
             $table->unsignedInteger('application_id');
-            $table->smallInteger('rate');
+            $table->tinyInteger('rate');
             $table->softDeletes();
             $table->timestamps();
         });
